feat: improve file controller
move file.py to base.py
make label optional
